Setup creationg of child tool
