+++ mmeds/database.py
@@ -1,10 +1,13 @@
-import pymysql as pms
-import mongoengine as men
-import cherrypy as cp
-import pandas as pd
 import os
 import shutil
 import warnings
+
+import mmeds.secrets as sec
+import mmeds.config as fig
+import mongoengine as men
+import pymysql as pms
+import cherrypy as cp
+import pandas as pd
 
 from datetime import datetime
 from pathlib import WindowsPath, Path
@@ -13,57 +16,9 @@ from collections import defaultdict
 from mmeds.config import TABLE_ORDER, MMEDS_EMAIL, USER_FILES, SQL_DATABASE, get_salt
 from mmeds.error import TableAccessError, MissingUploadError, MetaDataError, NoResultError
 from mmeds.util import send_email, log, pyformat_translate, quote_sql, parse_ICD_codes
-import mmeds.secrets as sec
-import mmeds.config as fig
+from mmeds.documents import MetaData
 
 DAYS = 13
-
-
-class MetaData(men.DynamicDocument):
-    created = men.DateTimeField()
-    last_accessed = men.DateTimeField()
-    study_type = men.StringField(max_length=45, required=True)
-    reads_type = men.StringField(max_length=45, required=True)
-    study = men.StringField(max_length=45, required=True)
-    access_code = men.StringField(max_length=50, required=True)
-    owner = men.StringField(max_length=100, required=True)
-    email = men.StringField(max_length=100, required=True)
-    path = men.StringField(max_length=100, required=True)
-    metadata = men.DictField()
-    files = men.DictField()
-
-    # When the document is updated record the
-    # location of all files in a new file
-    def save(self):
-        with open(str(Path(self.path) / 'file_index.tsv'), 'w') as f:
-            f.write('{}\t{}\t{}\n'.format(self.owner, self.email, self.access_code))
-            f.write('Key\tPath\n')
-            for key, file_path in self.files.items():
-                # Skip non existent files
-                if file_path is None:
-                    continue
-                # If it's a key for an analysis point to the file index for that analysis
-                elif isinstance(file_path, dict):
-                    f.write('{}\t{}\n'.format(key, Path(self.path) / key / 'file_index.tsv'))
-                # Otherwise just write the value
-                else:
-                    f.write('{}\t{}\n'.format(key, file_path))
-        super(MetaData, self).save()
-
-    def __str__(self):
-        self_string = 'Created: {created}\n last_accessed: {last_accessed}\n study_type: {study_type}\n' +\
-            'reads_type: {reads_type}\n study: {study}\n access_code: {access_code}\n owner: {owner}\n' +\
-            'email: {email}\n path: {path}\n'
-        self_string = self_string.format(created=self.created, last_accessed=self.last_accessed,
-                                         study_type=self.study_type, reads_type=self.reads_type,
-                                         study=self.study, access_code=self.access_code,
-                                         owner=self.owner, email=self.email, path=self.path)
-        self_string += 'files: {}\n'.format(self.files.keys())
-        return self_string
-
-
-class MongoConnection:
-    pass
 
 
 class Database:
